move campaign enum to core
