+++ apps/hmm1/src/app/Menu/CampaignMenu/index.ts
@@ -1,1 +1,1 @@
-export { Campaign, CampaignMenu } from './CampaignMenu';
+export { CampaignMenu } from './CampaignMenu';
